requirements.txt
addict==2.4.0
altair==5.3.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
branca==0.7.2
cachetools==5.3.3
Cartopy==0.23.0
certifi==2024.2.2
cffi==1.16.0
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dask==2024.4.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.14.0
fiona==1.9.6
folium==0.16.0
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.3.1
geopandas==0.14.4
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
h5netcdf==1.3.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
icoscp==0.1.20
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.25
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nc-time-axis==1.4.1
nest-asyncio==1.6.0
netCDF4==1.6.5
networkx==3.3
notebook_shim==0.2.4
numcodecs==0.12.1
numexpr==2.10.0
numpy==1.26.4
openghg==0.8.0
openghg-defs==0.0.2
openghg_calscales==0.1.1
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.0
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
scipy==1.13.0
Send2Trash==1.8.3
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.37.1
sympy==1.13.1
tenacity==8.2.3
terminado==0.18.1
tinycss2==1.3.0
tinydb==4.8.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
windrose==1.9.2
xarray==2024.3.0
xyzservices==2024.4.0
zarr==2.17.2
zipp==3.18.1
